requirements.txt
torch==2.1.1
torchvision==0.16.1
numpy==1.26.4
Pillow==12.0.0
tqdm==4.67.1
matplotlib==3.9.2
scikit-learn==1.3.0
scikit-image==0.25.2
pytest==9.0.2